сложение рандомных чисел

+++ Example004_Div/Program.cs
@@ -1,4 +1,3 @@
 ﻿double numberA = 12;
-double numberB = 5;
-Console.WriteLine (numberA / numberB);
-
+int numberB = 5;
+Console.WriteLine(numberA / numberB);
